fold spec.test2 into spec.test
